src/web_ui/controlpanel/requirements.txt
flask
flask-login
flask-sqlalchemy
dotenv
requests
pillow
pyzbar